@@ -24,9 +24,9 @@
   "devDependencies": {
     "@parcel/packager-raw-url": "^2.2.1",
     "@parcel/transformer-webmanifest": "^2.2.1",
-    "@types/react": "^16.9.34",
+    "@types/react": "^17.0.38",
     "@types/react-color": "^3.0.1",
-    "@types/react-dom": "^16.9.7",
+    "@types/react-dom": "^17.0.11",
     "@types/react-router-dom": "^5.1.5",
     "@types/uuid": "^7.0.3",
     "@typescript-eslint/eslint-plugin": "^2.30.0",
@@ -45,15 +45,14 @@
     "typescript": "^3.8.3"
   },
   "dependencies": {
-    "@material-ui/core": "^4.9.12",
-    "@material-ui/icons": "^4.9.1",
+    "@mui/icons-material": "^5.3.0",
+    "@mui/material": "^5.3.0",
     "dayjs": "^1.8.25",
-    "react": "^16.13.1",
+    "react": "^17.0.2",
     "react-color": "^2.18.0",
     "react-day-picker": "^7.4.8",
-    "react-dom": "^16.13.1",
+    "react-dom": "^17.0.2",
     "react-router-dom": "^5.1.2",
-    "styled-components": "^5.1.0",
     "uuid": "^8.0.0"
   }
 }
